requirements.txt
flask==2.3.3
flask-cors==4.0.0
opencv-python==4.8.1.78
ultralytics==8.0.0
deep-sort-realtime==1.3.2
torch==2.0.1
torchvision==0.15.2
numpy==1.24.3
pandas==2.0.3
plotly==5.15.0
Pillow==10.0.0
scipy==1.11.0